Add initial generic testing
